refactor: 개봉 에정 영화 주가 정보 fetch 안 되도록 수정
