+++ src/pages/MovieDetail/index.tsx
@@ -1,15 +1,13 @@
-import React, { useEffect } from 'react';
+import React, { Suspense, useEffect } from 'react';
 import styled from 'styled-components';
-import { Txt } from '@/components/common/Txt';
 import { Route, useParams } from 'react-router-dom';
 import { useGetMovieDetail } from '@/hooks/api/movie-detail/useGetMovieDetail';
 import { useNavigate } from 'react-router-dom';
-import Tooltip from '@/components/common/Tooltip';
 import { useGetPollBox } from '@/hooks/api/movie-detail/useGetPollBox';
-import { useGetStockInfo } from '@/hooks/api/movie-detail/useGetStockInfo';
 import MovieDetailAfterOpen from '@/components/movieDetail/MovieDetailAfterOpen';
 import MovieDetailBeforeOpen from '@/components/movieDetail/MovieDetailBeforeOpen';
 import { getTodayYYYYMMDD } from '@/utils/format';
+import MovieDetailPoster from '@/components/movieDetail/MovieDetailPoster';
 
 export default function MovieDetail() {
   const navigate = useNavigate();
@@ -25,11 +23,7 @@ export default function MovieDetail() {
 
   const { movieDetail } = useGetMovieDetail(movieId);
   const { pollBox } = useGetPollBox(movieId);
-  const { stockMovieInfo } = useGetStockInfo(movieId);
-  //   const { movieDetail } = useGetMovieDetail(movieId);
-  console.log(getTodayYYYYMMDD()); // 오늘 날짜 출력
-  console.log(movieDetail.movieOpenDate); // 영화 개봉일 출력
-  console.log(getTodayYYYYMMDD() < movieDetail.movieOpenDate); // 비교 결과 출력
+
   const handleUpVote = () => {
     console.log('오른다 선택됨');
   };
@@ -41,39 +35,19 @@ export default function MovieDetail() {
   return (
     <PageContainer>
       <div className="wrapper">
-        <PosterContainer>
-          <PosterImage
-            src={movieDetail.moviePoster?.[0]}
-            alt={movieDetail.movieTitle}
-          />
-          <Tooltip text={movieDetail.movieTitle} position="top">
-            <EllipsisTxt typography="Pretendard36bold">
-              {movieDetail.movieTitle}
-            </EllipsisTxt>
-          </Tooltip>
-          <Txt typography="Pretendard20bold" color="greyscale11">
-            {movieDetail.movieOpenDate} 개봉
-          </Txt>
-          <Txt typography="Pretendard20regular" color="greyscale10">
-            감독 {movieDetail.movieDirector}
-          </Txt>
-          <Txt typography="Pretendard20regular" color="greyscale10">
-            제작(배급사) {stockMovieInfo.companyName}
-          </Txt>
-        </PosterContainer>
+        <MovieDetailPoster movieDetail={movieDetail} />
 
         <RightContainer>
           {getTodayYYYYMMDD() < movieDetail.movieOpenDate ? (
             <MovieDetailBeforeOpen
               movieDetail={movieDetail}
               pollBox={pollBox}
-              // stockMovieInfo={stockMovieInfo}
             />
           ) : (
             <MovieDetailAfterOpen
               movieDetail={movieDetail}
               pollBox={pollBox}
-              stockMovieInfo={stockMovieInfo}
+              movieId={movieId}
             />
           )}
         </RightContainer>
@@ -94,34 +68,6 @@ const PageContainer = styled.div`
     width: 100%;
     align-items: center;
   }
-`;
-
-// 왼쪽 섹션 (포스터 및 기본 정보)
-const PosterContainer = styled.div`
-  /* width: 300px; */
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: center;
-  margin-right: 40px;
-`;
-
-const PosterImage = styled.img`
-  width: 400px;
-  height: 570px;
-  margin-bottom: 20px;
-`;
-
-const EllipsisTxt = styled(Txt)`
-  width: 400px;
-  margin-bottom: 10px;
-  text-overflow: ellipsis;
-  overflow: hidden;
-  word-break: break-word;
-
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
 `;
 
 const RightContainer = styled.div`
